v1.0

+++ src/app/shopping-list/shopping-list.component.ts
@@ -4,6 +4,7 @@ import { Subscription } from "rxjs";
 import { Product } from "../product/product.model";
 import { CartItem } from "../shared/cart-item.model";
 import { DataStorageService } from "../shared/data-storage.service";
+import { UserService } from "../shared/user.service";
 import { ShoppingListService } from "./shopping-list.service";
 
 @Component({
@@ -12,8 +13,13 @@ import { ShoppingListService } from "./shopping-list.service";
   styleUrls: ["./shopping-list.component.css"],
 })
 export class ShoppingListComponent implements OnInit {
-  constructor(private shoppingListService: ShoppingListService,private dataStorageService: DataStorageService) {}
+
+  loggedIn = false;
+  constructor(private shoppingListService: ShoppingListService,
+    private dataStorageService: DataStorageService,
+    private userService:UserService) {}
   subscription: Subscription;
+  loginSubscription : Subscription;
 
   products: CartItem[] = this.shoppingListService.getProducts();
 
@@ -21,16 +27,26 @@ export class ShoppingListComponent implements OnInit {
     this.subscription = this.shoppingListService.shoppingList.subscribe(() => {
       this.products = this.shoppingListService.getProducts();
     });
+    this.loginSubscription = this.userService.loggedInSubj.subscribe(() =>{
+      this.loggedIn = this.userService.loggedIn;
+    })
   }
 
   onDeleteFromBasket(index: number) {
     this.shoppingListService.deleteProduct(index);
   }
+  onDeleteItem(index:number){
+    this.shoppingListService.removeFromCart(index);
+  }
   onAddItem(index: number) {
     this.shoppingListService.ascendProduct(index);
   }
   onCheckout(){
-    this.dataStorageService.storeProducts();
-    this.shoppingListService.deleteProductsFromSL();
+    if (this.loggedIn) {
+      this.dataStorageService.storeProducts();
+      this.shoppingListService.deleteProductsFromSL();
+    } else {
+      console.log("please login");
+    }
   }
 }
